ipv6 addrconf: revert /proc/net/if_inet6 ifa_flag format
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-123.el7
Rebuild_CHGLOG: - [net] ipv6: addrconf: revert /proc/net/if_inet6 ifa_flag format (Jiri Pirko) [1021871]
Rebuild_FUZZ: 99.12%
commit-author Jiri Pirko <jiri@resnulli.us>
commit 971a351ccbbd2b6eef136a2221da0b80aca50906

Turned out that applications like ifconfig do not handle the change.
So revert ifa_flag format back to 2-letter hex value.

Introduced by:
commit 479840ffdbe4242e8a25349218c8e0859223aa35
    "ipv6 addrconf: extend ifa_flags to u32"

	Reported-by: Alexander Aring <alex.aring@gmail.com>
	Tested-by: FLorent Fourcot <florent.fourcot@enst-bretagne.fr>
(cherry picked from commit 971a351ccbbd2b6eef136a2221da0b80aca50906)

+++ net/ipv6/addrconf.c
@@ -3434,7 +3434,7 @@ static int if6_seq_show(struct seq_file *seq, void *v)
 		   ifp->idev->dev->ifindex,
 		   ifp->prefix_len,
 		   ifp->scope,
-		   ifp->flags,
+		   (u8) ifp->flags,
 		   ifp->idev->dev->name);
 	return 0;
 }
